Created Naive Eternal
This commit marks the start of something wonderful. So this is eternal.
An event parsing library that will parse text to ice_cube objects. The
current implementation is an extremely limited token parser. It will
look for keywords and build the ice_cube schedule based on that.
Eventually the goal is to maybe create an AST? Or look at an approach
similar to Facebook's duckling. For now the focus will be on simply
improving the rules into making eternal work.

This commit updates the gitignore to not include gemfile
Fixes gemspec ( was using wrong variable for definition )

Adds patterns, which are a number of regexes that could potentially
become useful for parsing schedules. It also adds the main eternal code
which deals with tokenizing, cleaning, parsing, and evaluating the text
it parses. It's basically a mini programming language implementation.
Eventually it may be beneficial to look into a lexer and parser ?